mask-rcnn/requirements.txt
numpy==2.2.5
torch==2.6.0
torchvision==0.21.0
pycocotools==2.0.8
pillow==11.2.1
opencv-python==4.11.0.86
